requirements.txt
PyTelegramBotAPI
requests
flask
dotenv
google-generativeai